move schema getter to sep file
